revert bogus warning-silencer work from 2009 release
undoing the bogus casts and long pointercasts which were added in
b3d1162b5d2da08da6805b9db948c189f9c32f11.
we now mute the warning from unsigned-vs-signed char* paramter
passing, instead of hiding the non-issue behind a cast, which in
turn could prevent real issues from getting overlooked.
likewise, the pointer-to-int casts were replaced with long, but
the proper type is (u)intptr_t (will be fixed in next commit).

+++ src/cpc/diskimage/dsk.c
@@ -61,7 +61,7 @@ int		Dsk_Validate(const unsigned char *pDiskImage, const unsigned long DiskImage
 					{
 						/* correct size image */
 						int	i;
-						DSKTRACKHEADER *pTrack = (DSKTRACKHEADER *)((long)pHeader + sizeof(DSKHEADER));
+						DSKTRACKHEADER *pTrack = (DSKTRACKHEADER *)((int)pHeader + sizeof(DSKHEADER));
 
 						for (i=0; i<TotalTracks; i++)
 						{
@@ -80,7 +80,7 @@ int		Dsk_Validate(const unsigned char *pDiskImage, const unsigned long DiskImage
 							if (thisTrackSize>TrackSize)
 								break;
 
-							pTrack = (DSKTRACKHEADER *)((long)pTrack + TrackSize);
+							pTrack = (DSKTRACKHEADER *)((int)pTrack + TrackSize);
 						}
 
 						if (i==TotalTracks)
